keep sigmoid value computation in fp32

+++ verl/workers/critic/dp_critic.py
@@ -120,7 +120,7 @@ class DataParallelPPOCritic(BasePPOCritic):
                 values = pad_input(values_rmpad, indices=indices, batch=batch, seqlen=seqlen).squeeze(-1)
                 values = values[:, -response_length - 1 : -1]
                 if self.value_head_activation == "sigmoid":
-                    values = torch.sigmoid(values)
+                    values = torch.sigmoid(values.float())
                 elif self.value_head_activation == "tanh":
                     values = torch.tanh(values)
             else:
@@ -138,7 +138,7 @@ class DataParallelPPOCritic(BasePPOCritic):
                     values = output.logits
                 values = values[:, -response_length - 1 : -1].squeeze(-1)
                 if self.value_head_activation == "sigmoid":
-                    values = torch.sigmoid(values)
+                    values = torch.sigmoid(values.float())
                 elif self.value_head_activation == "tanh":
                     values = torch.tanh(values)
             return values
